@@ -4,15 +4,16 @@
 
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             (unknown)
 // source: srvpb/v1/oracle.proto
 
 package srv
 
 import (
+	v1 "buf.build/gen/go/luthersystems/protos/protocolbuffers/go/healthcheck/v1"
 	context "context"
-	v1 "github.com/luthersystems/sandbox/api/pb/v1"
+	v11 "github.com/luthersystems/sandbox/api/pb/v1"
 	grpc "google.golang.org/grpc"
 	codes "google.golang.org/grpc/codes"
 	status "google.golang.org/grpc/status"
@@ -20,11 +21,11 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
-	LedgerService_HealthCheck_FullMethodName     = "/srvpb.v1.LedgerService/HealthCheck"
+	LedgerService_GetHealthCheck_FullMethodName  = "/srvpb.v1.LedgerService/GetHealthCheck"
 	LedgerService_CreateAccount_FullMethodName   = "/srvpb.v1.LedgerService/CreateAccount"
 	LedgerService_UpdateAccount_FullMethodName   = "/srvpb.v1.LedgerService/UpdateAccount"
 	LedgerService_DeleteAccount_FullMethodName   = "/srvpb.v1.LedgerService/DeleteAccount"
@@ -36,21 +37,23 @@ const (
 // LedgerServiceClient is the client API for LedgerService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// REST-JSON API Docs for the ledger project.
 type LedgerServiceClient interface {
 	// Check the health of the service. This is used by load balancers to check service health.
-	HealthCheck(ctx context.Context, in *v1.HealthCheckRequest, opts ...grpc.CallOption) (*v1.HealthCheckResponse, error)
+	GetHealthCheck(ctx context.Context, in *v1.GetHealthCheckRequest, opts ...grpc.CallOption) (*v1.GetHealthCheckResponse, error)
 	// Create a new account denominated in an underlying asset.
-	CreateAccount(ctx context.Context, in *v1.CreateAccountRequest, opts ...grpc.CallOption) (*v1.CreateAccountResponse, error)
+	CreateAccount(ctx context.Context, in *v11.CreateAccountRequest, opts ...grpc.CallOption) (*v11.CreateAccountResponse, error)
 	// Update an account.
-	UpdateAccount(ctx context.Context, in *v1.UpdateAccountRequest, opts ...grpc.CallOption) (*v1.UpdateAccountResponse, error)
+	UpdateAccount(ctx context.Context, in *v11.UpdateAccountRequest, opts ...grpc.CallOption) (*v11.UpdateAccountResponse, error)
 	// Delete an account.
-	DeleteAccount(ctx context.Context, in *v1.DeleteAccountRequest, opts ...grpc.CallOption) (*v1.DeleteAccountResponse, error)
+	DeleteAccount(ctx context.Context, in *v11.DeleteAccountRequest, opts ...grpc.CallOption) (*v11.DeleteAccountResponse, error)
 	// Retrieve account details.
-	GetAccount(ctx context.Context, in *v1.GetAccountRequest, opts ...grpc.CallOption) (*v1.GetAccountResponse, error)
+	GetAccount(ctx context.Context, in *v11.GetAccountRequest, opts ...grpc.CallOption) (*v11.GetAccountResponse, error)
 	// Get accounts for a user.
-	GetUserAccounts(ctx context.Context, in *v1.GetUserAccountsRequest, opts ...grpc.CallOption) (*v1.GetUserAccountsResponse, error)
+	GetUserAccounts(ctx context.Context, in *v11.GetUserAccountsRequest, opts ...grpc.CallOption) (*v11.GetUserAccountsResponse, error)
 	// Transfer between accounts denominated in the same underlying asset.
-	Transfer(ctx context.Context, in *v1.TransferRequest, opts ...grpc.CallOption) (*v1.TransferResponse, error)
+	Transfer(ctx context.Context, in *v11.TransferRequest, opts ...grpc.CallOption) (*v11.TransferResponse, error)
 }
 
 type ledgerServiceClient struct {
@@ -61,63 +64,70 @@ func NewLedgerServiceClient(cc grpc.ClientConnInterface) LedgerServiceClient {
 	return &ledgerServiceClient{cc}
 }
 
-func (c *ledgerServiceClient) HealthCheck(ctx context.Context, in *v1.HealthCheckRequest, opts ...grpc.CallOption) (*v1.HealthCheckResponse, error) {
-	out := new(v1.HealthCheckResponse)
-	err := c.cc.Invoke(ctx, LedgerService_HealthCheck_FullMethodName, in, out, opts...)
+func (c *ledgerServiceClient) GetHealthCheck(ctx context.Context, in *v1.GetHealthCheckRequest, opts ...grpc.CallOption) (*v1.GetHealthCheckResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(v1.GetHealthCheckResponse)
+	err := c.cc.Invoke(ctx, LedgerService_GetHealthCheck_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *ledgerServiceClient) CreateAccount(ctx context.Context, in *v1.CreateAccountRequest, opts ...grpc.CallOption) (*v1.CreateAccountResponse, error) {
-	out := new(v1.CreateAccountResponse)
-	err := c.cc.Invoke(ctx, LedgerService_CreateAccount_FullMethodName, in, out, opts...)
+func (c *ledgerServiceClient) CreateAccount(ctx context.Context, in *v11.CreateAccountRequest, opts ...grpc.CallOption) (*v11.CreateAccountResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(v11.CreateAccountResponse)
+	err := c.cc.Invoke(ctx, LedgerService_CreateAccount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *ledgerServiceClient) UpdateAccount(ctx context.Context, in *v1.UpdateAccountRequest, opts ...grpc.CallOption) (*v1.UpdateAccountResponse, error) {
-	out := new(v1.UpdateAccountResponse)
-	err := c.cc.Invoke(ctx, LedgerService_UpdateAccount_FullMethodName, in, out, opts...)
+func (c *ledgerServiceClient) UpdateAccount(ctx context.Context, in *v11.UpdateAccountRequest, opts ...grpc.CallOption) (*v11.UpdateAccountResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(v11.UpdateAccountResponse)
+	err := c.cc.Invoke(ctx, LedgerService_UpdateAccount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *ledgerServiceClient) DeleteAccount(ctx context.Context, in *v1.DeleteAccountRequest, opts ...grpc.CallOption) (*v1.DeleteAccountResponse, error) {
-	out := new(v1.DeleteAccountResponse)
-	err := c.cc.Invoke(ctx, LedgerService_DeleteAccount_FullMethodName, in, out, opts...)
+func (c *ledgerServiceClient) DeleteAccount(ctx context.Context, in *v11.DeleteAccountRequest, opts ...grpc.CallOption) (*v11.DeleteAccountResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(v11.DeleteAccountResponse)
+	err := c.cc.Invoke(ctx, LedgerService_DeleteAccount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *ledgerServiceClient) GetAccount(ctx context.Context, in *v1.GetAccountRequest, opts ...grpc.CallOption) (*v1.GetAccountResponse, error) {
-	out := new(v1.GetAccountResponse)
-	err := c.cc.Invoke(ctx, LedgerService_GetAccount_FullMethodName, in, out, opts...)
+func (c *ledgerServiceClient) GetAccount(ctx context.Context, in *v11.GetAccountRequest, opts ...grpc.CallOption) (*v11.GetAccountResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(v11.GetAccountResponse)
+	err := c.cc.Invoke(ctx, LedgerService_GetAccount_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *ledgerServiceClient) GetUserAccounts(ctx context.Context, in *v1.GetUserAccountsRequest, opts ...grpc.CallOption) (*v1.GetUserAccountsResponse, error) {
-	out := new(v1.GetUserAccountsResponse)
-	err := c.cc.Invoke(ctx, LedgerService_GetUserAccounts_FullMethodName, in, out, opts...)
+func (c *ledgerServiceClient) GetUserAccounts(ctx context.Context, in *v11.GetUserAccountsRequest, opts ...grpc.CallOption) (*v11.GetUserAccountsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(v11.GetUserAccountsResponse)
+	err := c.cc.Invoke(ctx, LedgerService_GetUserAccounts_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *ledgerServiceClient) Transfer(ctx context.Context, in *v1.TransferRequest, opts ...grpc.CallOption) (*v1.TransferResponse, error) {
-	out := new(v1.TransferResponse)
-	err := c.cc.Invoke(ctx, LedgerService_Transfer_FullMethodName, in, out, opts...)
+func (c *ledgerServiceClient) Transfer(ctx context.Context, in *v11.TransferRequest, opts ...grpc.CallOption) (*v11.TransferResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(v11.TransferResponse)
+	err := c.cc.Invoke(ctx, LedgerService_Transfer_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -126,51 +136,57 @@ func (c *ledgerServiceClient) Transfer(ctx context.Context, in *v1.TransferReque
 
 // LedgerServiceServer is the server API for LedgerService service.
 // All implementations must embed UnimplementedLedgerServiceServer
-// for forward compatibility
+// for forward compatibility.
+//
+// REST-JSON API Docs for the ledger project.
 type LedgerServiceServer interface {
 	// Check the health of the service. This is used by load balancers to check service health.
-	HealthCheck(context.Context, *v1.HealthCheckRequest) (*v1.HealthCheckResponse, error)
+	GetHealthCheck(context.Context, *v1.GetHealthCheckRequest) (*v1.GetHealthCheckResponse, error)
 	// Create a new account denominated in an underlying asset.
-	CreateAccount(context.Context, *v1.CreateAccountRequest) (*v1.CreateAccountResponse, error)
+	CreateAccount(context.Context, *v11.CreateAccountRequest) (*v11.CreateAccountResponse, error)
 	// Update an account.
-	UpdateAccount(context.Context, *v1.UpdateAccountRequest) (*v1.UpdateAccountResponse, error)
+	UpdateAccount(context.Context, *v11.UpdateAccountRequest) (*v11.UpdateAccountResponse, error)
 	// Delete an account.
-	DeleteAccount(context.Context, *v1.DeleteAccountRequest) (*v1.DeleteAccountResponse, error)
+	DeleteAccount(context.Context, *v11.DeleteAccountRequest) (*v11.DeleteAccountResponse, error)
 	// Retrieve account details.
-	GetAccount(context.Context, *v1.GetAccountRequest) (*v1.GetAccountResponse, error)
+	GetAccount(context.Context, *v11.GetAccountRequest) (*v11.GetAccountResponse, error)
 	// Get accounts for a user.
-	GetUserAccounts(context.Context, *v1.GetUserAccountsRequest) (*v1.GetUserAccountsResponse, error)
+	GetUserAccounts(context.Context, *v11.GetUserAccountsRequest) (*v11.GetUserAccountsResponse, error)
 	// Transfer between accounts denominated in the same underlying asset.
-	Transfer(context.Context, *v1.TransferRequest) (*v1.TransferResponse, error)
+	Transfer(context.Context, *v11.TransferRequest) (*v11.TransferResponse, error)
 	mustEmbedUnimplementedLedgerServiceServer()
 }
 
-// UnimplementedLedgerServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedLedgerServiceServer struct {
-}
+// UnimplementedLedgerServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedLedgerServiceServer struct{}
 
-func (UnimplementedLedgerServiceServer) HealthCheck(context.Context, *v1.HealthCheckRequest) (*v1.HealthCheckResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method HealthCheck not implemented")
+func (UnimplementedLedgerServiceServer) GetHealthCheck(context.Context, *v1.GetHealthCheckRequest) (*v1.GetHealthCheckResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetHealthCheck not implemented")
 }
-func (UnimplementedLedgerServiceServer) CreateAccount(context.Context, *v1.CreateAccountRequest) (*v1.CreateAccountResponse, error) {
+func (UnimplementedLedgerServiceServer) CreateAccount(context.Context, *v11.CreateAccountRequest) (*v11.CreateAccountResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateAccount not implemented")
 }
-func (UnimplementedLedgerServiceServer) UpdateAccount(context.Context, *v1.UpdateAccountRequest) (*v1.UpdateAccountResponse, error) {
+func (UnimplementedLedgerServiceServer) UpdateAccount(context.Context, *v11.UpdateAccountRequest) (*v11.UpdateAccountResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateAccount not implemented")
 }
-func (UnimplementedLedgerServiceServer) DeleteAccount(context.Context, *v1.DeleteAccountRequest) (*v1.DeleteAccountResponse, error) {
+func (UnimplementedLedgerServiceServer) DeleteAccount(context.Context, *v11.DeleteAccountRequest) (*v11.DeleteAccountResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method DeleteAccount not implemented")
 }
-func (UnimplementedLedgerServiceServer) GetAccount(context.Context, *v1.GetAccountRequest) (*v1.GetAccountResponse, error) {
+func (UnimplementedLedgerServiceServer) GetAccount(context.Context, *v11.GetAccountRequest) (*v11.GetAccountResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetAccount not implemented")
 }
-func (UnimplementedLedgerServiceServer) GetUserAccounts(context.Context, *v1.GetUserAccountsRequest) (*v1.GetUserAccountsResponse, error) {
+func (UnimplementedLedgerServiceServer) GetUserAccounts(context.Context, *v11.GetUserAccountsRequest) (*v11.GetUserAccountsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetUserAccounts not implemented")
 }
-func (UnimplementedLedgerServiceServer) Transfer(context.Context, *v1.TransferRequest) (*v1.TransferResponse, error) {
+func (UnimplementedLedgerServiceServer) Transfer(context.Context, *v11.TransferRequest) (*v11.TransferResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Transfer not implemented")
 }
 func (UnimplementedLedgerServiceServer) mustEmbedUnimplementedLedgerServiceServer() {}
+func (UnimplementedLedgerServiceServer) testEmbeddedByValue()                       {}
 
 // UnsafeLedgerServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to LedgerServiceServer will
@@ -180,29 +196,36 @@ type UnsafeLedgerServiceServer interface {
 }
 
 func RegisterLedgerServiceServer(s grpc.ServiceRegistrar, srv LedgerServiceServer) {
+	// If the following call pancis, it indicates UnimplementedLedgerServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&LedgerService_ServiceDesc, srv)
 }
 
-func _LedgerService_HealthCheck_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(v1.HealthCheckRequest)
+func _LedgerService_GetHealthCheck_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(v1.GetHealthCheckRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(LedgerServiceServer).HealthCheck(ctx, in)
+		return srv.(LedgerServiceServer).GetHealthCheck(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: LedgerService_HealthCheck_FullMethodName,
+		FullMethod: LedgerService_GetHealthCheck_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LedgerServiceServer).HealthCheck(ctx, req.(*v1.HealthCheckRequest))
+		return srv.(LedgerServiceServer).GetHealthCheck(ctx, req.(*v1.GetHealthCheckRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
 func _LedgerService_CreateAccount_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(v1.CreateAccountRequest)
+	in := new(v11.CreateAccountRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -214,13 +237,13 @@ func _LedgerService_CreateAccount_Handler(srv interface{}, ctx context.Context, 
 		FullMethod: LedgerService_CreateAccount_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LedgerServiceServer).CreateAccount(ctx, req.(*v1.CreateAccountRequest))
+		return srv.(LedgerServiceServer).CreateAccount(ctx, req.(*v11.CreateAccountRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
 func _LedgerService_UpdateAccount_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(v1.UpdateAccountRequest)
+	in := new(v11.UpdateAccountRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -232,13 +255,13 @@ func _LedgerService_UpdateAccount_Handler(srv interface{}, ctx context.Context, 
 		FullMethod: LedgerService_UpdateAccount_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LedgerServiceServer).UpdateAccount(ctx, req.(*v1.UpdateAccountRequest))
+		return srv.(LedgerServiceServer).UpdateAccount(ctx, req.(*v11.UpdateAccountRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
 func _LedgerService_DeleteAccount_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(v1.DeleteAccountRequest)
+	in := new(v11.DeleteAccountRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -250,13 +273,13 @@ func _LedgerService_DeleteAccount_Handler(srv interface{}, ctx context.Context, 
 		FullMethod: LedgerService_DeleteAccount_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LedgerServiceServer).DeleteAccount(ctx, req.(*v1.DeleteAccountRequest))
+		return srv.(LedgerServiceServer).DeleteAccount(ctx, req.(*v11.DeleteAccountRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
 func _LedgerService_GetAccount_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(v1.GetAccountRequest)
+	in := new(v11.GetAccountRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -268,13 +291,13 @@ func _LedgerService_GetAccount_Handler(srv interface{}, ctx context.Context, dec
 		FullMethod: LedgerService_GetAccount_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LedgerServiceServer).GetAccount(ctx, req.(*v1.GetAccountRequest))
+		return srv.(LedgerServiceServer).GetAccount(ctx, req.(*v11.GetAccountRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
 func _LedgerService_GetUserAccounts_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(v1.GetUserAccountsRequest)
+	in := new(v11.GetUserAccountsRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -286,13 +309,13 @@ func _LedgerService_GetUserAccounts_Handler(srv interface{}, ctx context.Context
 		FullMethod: LedgerService_GetUserAccounts_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LedgerServiceServer).GetUserAccounts(ctx, req.(*v1.GetUserAccountsRequest))
+		return srv.(LedgerServiceServer).GetUserAccounts(ctx, req.(*v11.GetUserAccountsRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
 func _LedgerService_Transfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(v1.TransferRequest)
+	in := new(v11.TransferRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -304,7 +327,7 @@ func _LedgerService_Transfer_Handler(srv interface{}, ctx context.Context, dec f
 		FullMethod: LedgerService_Transfer_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LedgerServiceServer).Transfer(ctx, req.(*v1.TransferRequest))
+		return srv.(LedgerServiceServer).Transfer(ctx, req.(*v11.TransferRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -317,8 +340,8 @@ var LedgerService_ServiceDesc = grpc.ServiceDesc{
 	HandlerType: (*LedgerServiceServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
-			MethodName: "HealthCheck",
-			Handler:    _LedgerService_HealthCheck_Handler,
+			MethodName: "GetHealthCheck",
+			Handler:    _LedgerService_GetHealthCheck_Handler,
 		},
 		{
 			MethodName: "CreateAccount",
